+++ Project1/Project1/Utils.c
@@ -10,7 +10,7 @@ All content © 2021 DigiPen Institute of Technology Singapore, all rights reserve
 CP_Vector AngleToVector(float angle)
 {
 	float rads = CP_Math_Radians(angle);
-	CP_Vector ret = { (float) cos(rads),(float) sin(rads) };
+	CP_Vector ret = { cos(rads),sin(rads) };
 
 	return CP_Vector_Normalize(ret);
 }
